requirements.txt
selenium==3.141.0
pytest==6.2.2
allure-pytest==2.8.34
Pillow==8.1.0
pytest-dependency==0.5.1